requirements.txt
torch>=1.9.0
torchtext>=0.10.0
pytorch-lightning>=1.3.5
antlr4-python3-runtime==4.8
setuptools>=52.0.0
tqdm==4.58.0
numpy>=1.20.1
regex>=2019.11.1
omegaconf~=2.0.6
dataclasses~=0.6
requests~=2.25.1
pandas~=1.1.3
pycode2seq~=0.0.6
scikit-learn~=0.24.1
torchmetrics~=0.5.1
catboost~=0.25
pandas~=1.3.4
pytorch-crf~=0.7.2